vision-service/requirements.txt
flask>=2.0.0
flask-cors>=4.0.0
Pillow>=9.0.0
# OCR - install one:
pytesseract>=0.3.10
# Optional fallback: easyocr opencv-python
# YOLO: ultralytics torch torchvision
